requirements.txt
asgiref==3.4.1
bootstrap4==0.1.0
cffi==1.15.0
cryptography==35.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.13.0
django-import-export==2.6.1
django-rest-knox==4.1.0
django-viewflow==2.0.0a0
djangorestframework==3.12.4
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
pycparser==2.21
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
xlrd==2.0.1
xlwt==1.3.0
